5.1.33      *  GTK alpha version.      *  Abolished resolution independence: the Sound window now looks the        same on all platforms, and the Demo window has the same relative        font size on your screen and on the video projector.      *  GTK: support for asynchronous audio output.      *  GTK: sound plays once rather than three times in Sound and other        windows.      *  GTK: can click more than once in the manual.      *  GTK: correct pink selections in Sound and other windows.      *  GTK: correct dragging in TextGrid, Manipulation, and tier        windows.      *  GTK: a working TextGrid window.      *  GTK: no automatic triple click in the manual (and many other        windows).      *  GTK: moving cursor while sound is playing in Sound and other        windows.      *  GTK: correct colours of the rectangles in Sound and other        windows.      *  GTK: a working Group button.      *  GTK: correct font menus, font size menus, colour menus, line type        menus.      *  GTK: scrolling in the manual (and many other windows).      *  GTK: erase old texts in manual windows.      *  GTK: made Picture window come back when drawing.      *  GTK: Info window is fronted automatically.      *  GTK: support sendpraat.

+++ sys/melder_textencoding.c
@@ -286,7 +286,7 @@ int Melder_8bitToWcs_inline (const char *string, wchar_t *wcs, int inputEncoding
 	}
 	if (inputEncoding == kMelder_textInputEncoding_UTF8) {
 		while (*p != '\0') {
-			unsigned long kar = * p ++;
+			uint32_t kar = * p ++;
 			if (kar <= 0x7F) {
 				* q ++ = kar;
 			} else if (kar <= 0xDF) {
